+++ index.html
@@ -42,7 +42,7 @@ function main() {
     <div id="parametersdiv"></div>
     <div id="selectdiv"></div>
 
-    <div id="tail">
+    <div id="tail" style="position: absolute; bottom: 0;">
       <div id="statusdiv"></div>
     </div>
   </body>
