+++ src/io/generateOutputFileBlobUrl.js
@@ -1,6 +1,3 @@
-import { isSafari } from '../ui/detectBrowser'
-import { getWindowURL } from '../ui/urlHelpers'
-
 export default function generateOutputFileBlobUrl (extension, blob, callback) {
   if (isSafari()) {
     // console.log("Trying download via DATA URI")
@@ -20,4 +17,15 @@ export default function generateOutputFileBlobUrl (extension, blob, callback) {
     if (!outputFileBlobUrl) throw new Error('createObjectURL() failed')
     callback(outputFileBlobUrl, 'openjscad.' + extension, true, false)
   }
+}
+
+export function getWindowURL () {
+  if (window.URL) return window.URL
+  else if (window.webkitURL) return window.webkitURL
+  else throw new Error("Your browser doesn't support window.URL")
+}
+
+export function isSafari () {
+  // FIXME WWW says don't use this
+  return /Version\/[\d\.]+.*Safari/.test(window.navigator.userAgent)
 }
